+++ CHANGELOG.md
@@ -18,7 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Warn when labels precede captions or first numbered items.
 - Check section heading hierarchy.
 - Add a check for dash length in ranges and compounds.
-- Support selecting disabled rules, with `MAT004` disabled by default.
+- Support selecting disabled rules.
+- Support selecting rules from the command line.
 - More casing checks.
 
 ### Fixed